+++ core/app-version.js
@@ -1,1 +1,1 @@
-export const APP_VERSION = '20260706-6';
+export const APP_VERSION = '20260707-1';