+++ modules/graph-assumption-builder.js
@@ -1,10 +1,10 @@
-import { GraphUtils } from './graph-utils.js?v=20260706-6';
+import { GraphUtils } from './graph-utils.js?v=20260707-1';
 import {
     buildAssumptionNodeTooltip,
     buildAttackEdgeTooltip,
     buildJunctionTooltip,
     buildTopNodeTooltip
-} from './graph-tooltip-builder.js?v=20260706-6';
+} from './graph-tooltip-builder.js?v=20260707-1';
 
 // Cap on the number of minimal support sets enumerated per contrary, guarding
 // against combinatorial blow-up on pathological (deeply disjunctive) frameworks.